jdbc obfuscation
